Move ConfigBucketStack and CoreStacks to separate package
