+++ java/cdk/src/main/java/sleeper/cdk/stack/CompactionStack.java
@@ -76,6 +76,7 @@ import software.constructs.IDependable;
 
 import sleeper.cdk.jars.BuiltJars;
 import sleeper.cdk.jars.LambdaCode;
+import sleeper.cdk.stack.core.CoreStacks;
 import sleeper.cdk.util.Utils;
 import sleeper.configuration.CompactionTaskRequirements;
 import sleeper.core.ContainerConstants;
